Adding Coordinate Spaces
CoordinateSpace defines an abstract class which has functions to go back and forth between world space and a coordinate space.
CCFSpace shows an example of how this can be implemented, or see AllenCCFSpace25 for the same example but in 25 um unit space.

These functions have not been refactored into Pinpoint yet